Remove cd override for envrc, md func fix
The gopath function used by the cd override was conflicting with envrc
configurations.  Going with envrc as it's more customizable and can be
more easily shared.  Fix to md func to correct 'bad floating point
constant' error that seems to randomly pop up.  If this doesn't address
try wrapping $i in double quotes.