Clean header, use library, make links more robust
Updating the header to be consistent with other scripts in this repo.
Making use of the recently broken out library.sh to reduce code
duplication. Improving the symlinking function to be a bit more robust
with paths containing spaces. Setting the script to make paths and link
from the users home directory in order to simplify configs.