Rework repo completion for quicker navigation
Instead of having to use completion for the full path of the
repositories (each nested in subdirectories of the main repo folder,
grouped by service url, group, user, etc) search recursively for all
repos, present repo names as completion options, and allow those to be
enough to cd you to the right location. Like zsh cdpath but dynamic and
wrapped behind a `repo` command.