Fix fold function to reduce calls
The return '=' appears to kick of a recursive call of the fold
expression so adding back in the conditional above it more often than
not catches all remaining cases. It drops calls to the fold expression
from 66K times on save to 1K times on a test markdown file with ~200
lines. Consequently save times go from 3 to 0.03 seconds.