Refactor hammerspoon, get media keys to work
OSX doesn't let you pause your music with the media keys if you have
handoff turned on and you are receiving a phone call. So incredibly
frustrating. Fix was to use hammerspoon to override the media keys and
have them ONLY control itunes/music and do so directly, bypassing
whatever dumb logic goes on behind the scenes to overload those keys.
That had stopped working, any hs.grid call seems to break it. Fix is to
have the event watcher called in an imported module. This returns
consistent functioning.