Bulk update neovim plugins
Moving more configs to be self contained in after/plugins/* and only be
executed if the plugin is active. Wanting to consolidate things down and
make them more modular, ie be able to turn off one plugin and not have
it break anything or have any remnants other than the config file.