Smart add to prevent duplicate code pasting
If you call CodeCompanion with a visual selection it does the same thing
as CodeCompanion Add if it is creating a new chat buffer. So if you have
no existing chat buffers and call CodeCompanion Add you get two entries
of the code into the new chat buffer. This function handles that
scenario.