Improve smart_chat_add, add smart_inline
Improved workflows for improving common contexts. Lets you run an inline
chat start with no selection and assume the current buffer. Falls back
to default behavior when starting an inline chat with a visual
selection. Improve smart_chat_add to work with both codecompanion
setting options for whether or not to start a chat buffer in insert mode
or not.