Temp disable diffs to auto accept code changes
Will be able to see the git diffs and accept or reject from there.
Otherwise it is too clunky when you are waiting for the agent to make
multiple changes, having to make sure you accept each before they
timeout, breaking the flow and ending up in a half finished state.