Add in editorconfig guards
This may... maybe... a fix for nvim freezing when in insert mode.
Something is causing it to go into an infinite loop and hang at random
points when editing buffers. Since putting this in I've not bumped into
the issue. It was a suggestion from ChatGPT and guards against
editorconfig being enabled for no name buffers. Leaving it for now.